tests: spin up a whole engine for unit tests
This way we can ensure we configure devices correctly, and we
can use cloud sync to get to credentials.
